renamed and moved files. Make inputs path general and not for demography component only
