+++ tests/test_wpp_builder.py
@@ -20,7 +20,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from pipeline.components import utils
+from pipeline.components.framework import utils
 from pipeline.components.demography import wpp
